// src/app/model/client.ts
export class Client {
  Id: number;
  ClientCode: string;
  ClientName: string;
  AliasName: string;
  ClientReferenceID: number;
  Industry: string;
  IndustryName: string;
  Status: number;
  LastUpdatedOn: Date;
  LastUpdatedBy: string;
  LocationId: number;
  TallyReferenceId: string;
  TransactionDate: string;
  CollectionAmount: number;
  GroupCompany: string;
  SignedInLocation: number;
  LineOfBusiness: string;
  AssignmentOneLiner: string;
  AssignmentDetails: string;
  SignUpDate: Date;
  SignUpDateStr: string;
  SignedBy: number;
  SignedByUser: string;
}